lockup detector: Kconfig fixes to seperate hard and soft lockup options
Patchwork-id: 32573
O-Subject: [RHEL6 PATCH 5/6] lockup detector: Kconfig fixes to seperate hard and
	soft lockup options
Bugzilla: 669808
RH-Acked-by: Peter Zijlstra <pzijlstr@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=669808

The ability to enable LOCKUP_DETECTOR but disable HARDLOCKUP_DETECTOR was
lost somewhere.  This change allows that to happen.  This needs to be pushed
upstream.

Also add a check for the softlockup Kconfig option.  This prevents s390
from compiling it.  Again needs to be pushed upstream.

Discovered during brew testing.

Signed-off-by: Don Zickus <dzickus@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>