netfilter: audit target to record accepted/dropped packets
Patchwork-id: 32579
O-Subject: [RHEL6.1 PATCH] netfilter: audit target to record accepted/dropped
	packets (BZ642391)
Bugzilla: 642391
RH-Acked-by: David S. Miller <davem@redhat.com>

This patch adds a new netfilter target which creates audit records
for packets traversing a certain chain.

This functionality is required for the Common Criteria certification
for virtualization. (See BZ for more information)

It can be used to record packets which are rejected administraively
as follows:

  -N AUDIT_DROP
  -A AUDIT_DROP -j AUDIT --type DROP
  -A AUDIT_DROP -j DROP

a rule which would typically drop or reject a packet would then
invoke the new chain to record packets before dropping them.

  -j AUDIT_DROP

The module is protocol independant and works for iptables, ip6tables
and ebtables.

The following information is logged:
 - netfilter hook
 - packet length
 - incomming/outgoing interface
 - MAC src/dst/proto for ethernet packets
 - src/dst/protocol address for IPv4/IPv6
 - src/dst port for TCP/UDP/UDPLITE
 - icmp type/code

Upstream:
 - Patch was submitted upstream and is currently in its 3rd
   revision which I am confident is the last revision. The
   patch has not been merged yet though.

   http://marc.info/?l=netfilter-devel&m=129501842729593&w=2

Brew:
   https://brewweb.devel.redhat.com/taskinfo?taskID=3015784

Testing:
   Tested by myself and Thomas Woerner

Bugzilla:
  BZ642391

Signed-off-by: Thomas Graf <tgraf@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>