nfs4: set source address when callback is generated
Patchwork-id: 32586
O-Subject: [RHEL6.1 PATCH] nfs4: set source address when callback is generated
Bugzilla: 662589
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: Takuma Umeya <tumeya@redhat.com>

bz 662589

when callback is generated in NFSv4 server, it doesn't set the source
address. When an alias IP is utilized on NFSv4 server and suppose the
client is accessing via that alias IP (e.g. eth0:0), the client invokes
the callback to the IP address that is set on the original device (e.g.
eth0). This behavior results in timeout of xprt.
The patch sets the IP address that the client should invoke callback to.

Signed-off-by: Takuma Umeya <tumeya@redhat.com>
[bfields@redhat.com: Simplify gen_callback arguments, use helper function]
Signed-off-by: J. Bruce Fields <bfields@redhat.com>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>