inotify: rework inotify locking to prevent double free use when free in inotify
Patchwork-id: 32677
O-Subject: [PATCH] inotify: rework inotify locking to prevent double free use
	when free in inotify
Bugzilla: 582109

RH-BZ: 582109

If inotify watches are created a destroyed in highly multithreaded operations
it is possible that a mark will be used after freed or that mark may be double
freed.  This patch largely reworks the inotify mark creation and destruction
(especially the adding an removing from the idr) to match upstream, to remove
such races.

Signed-off-by: Eric Paris <eparis@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>