move kernel/unsupported.c to kernel/rh_taint.c
Patchwork-id: 32752
O-Subject: [RFC RHEL6 PATCH 1/2] misc: move kernel/unsupported.c to
	kernel/rh_taint.c
Bugzilla: 645198
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=645198

I was looking to expand the use of this file for more kernel
tainting.  It doesn't make sense to keep it as unsupported.c, move
it to something more relevant like rh_taint.c

The next patch adds to it.

Signed-off-by: Don Zickus <dzickus@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>