mptsas: Added-MPI_SCSIIO_CONTROL_HEADOFQ-priority
Patchwork-id: 32760
O-Subject: [RHEL6.1 PATCH 01/32] mptsas:
	Added-MPI_SCSIIO_CONTROL_HEADOFQ-priority
Bugzilla: 642618
RH-Acked-by: Rob Evers <revers@redhat.com>

There is a 'ioprio' field in the BIO and the Request structure.
check this priority field and set MPI_SCSIIO_CONTROL_HEADOFQ
to pass down I/O priority.
An enhancement to the LSI Disk Array Controller firmware is being
developed to look at the Head Of Queue bit to allow I/Os with the HOQ bit
set to be processed before I/Os which do not have the HOQ bit set.
In order to set the HOQ bit, the mpt fusion driver  needs to look at the
'ioprio' field in the request structure associated with the scsi command.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>