mptsas: Incorrect return value in mptscsih_dev_reset
Patchwork-id: 32787
O-Subject: [RHEL6.1 PATCH 31/32] mptsas: Incorrect return value in
	mptscsih_dev_reset
Bugzilla: 642618
RH-Acked-by: Rob Evers <revers@redhat.com>

There's a branch at the end of this function that
is supposed to normalize the return value with what
the mid-layer expects. In this one case, we get it wrong.

Also increase the verbosity of the INFO level printk
at the end of mptscsih_abort to include the actual return value
and the scmd->serial_number. The reason being success
or failure is actually determined by the state of
the internal tag list when a TMF is issued, and not the
return value of the TMF cmd. The serial_number is also
used in this decision, thus it's useful to know for debugging
purposes.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>