GFS2: [RFE] glock scalability patches
Patchwork-id: 32794
O-Subject: [RHEL6.1][GFS2]: Bug 656939 - GFS2: [RFE] glock scalability patches
Bugzilla: 656939
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

This is basically a performance change. The idea is to use rcu for lookups of glocks in the hash table (hence the need for hlist_bl_rcu) but at the same time it also reduces the code size (ignoring the new hlist_bl_rcu header) since the locking is also simplified as well.

This is already upstream.

Resolves: rhbz#656939
Signed-off-by: Steve Whitehouse <swhiteho@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>