ext4: serialize unaligned direct asynchronous IO
Patchwork-id: 32875
O-Subject: [PATCH RHEL6 V2] ext4: serialize unaligned direct asynchronous IO
Bugzilla: 615309
RH-Acked-by: Edward Shishkin <edward@redhat.com>

This is for Bug 615309,
qemu aio=native causes filesystem corruption with sparse images on EXT4

This patch has been submitted and initially acked upstream;
upstream re-used existing waitqueues, but I had to add that here:

ext4 has a data corruption case when doing non-block-aligned
asynchronous direct IO into a sparse file, as demonstrated
by xfstest 240.

The root cause is that while ext4 preallocates space in the
hole, mappings of that space still look "new" and
dio_zero_block() will zero out the unwritten portions.  When
more than one AIO thread is going, they both find this "new"
block and race to zero out their portion; this is uncoordinated
and causes data corruption.

Dave Chinner fixed this for xfs by simply serializing all
unaligned asynchronous direct IO.  I've done the same here.
The difference is that we only wait on conversions, not all IO.
This is a very big hammer, and I'm not very pleased with
stuffing this into ext4_file_write().  But since ext4 is
DIO_LOCKING, we need to serialize it at this high level.

I tried to move this into ext4_ext_direct_IO, but by then
we have the i_mutex already, and we will wait on the
work queue to do conversions - which must also take the
i_mutex.  So that won't work.

This was originally exposed by qemu-kvm installing to
a raw disk image with a normal sector-63 alignment.  I've
tested a backport of this patch with qemu, and it does
avoid the corruption.  It is also quite a lot slower
(14 min for package installs, vs. 8 min for well-aligned)
but I'll take slow correctness over fast corruption any day.

Mingming suggested that we can track outstanding
conversions, and wait on those so that non-sparse
files won't be affected, and I've implemented that here;
unaligned AIO to nonsparse files won't take a significant
perf hit (though AIO -submissions- are still serialized).

Signed-off-by: Eric Sandeen <sandeen@redhat.com>