Add new IRQ flag IRQF_NO_SUSPEND
Patchwork-id: 32878
O-Subject: [RHEL 6.1 PATCH 1/5 v2] irq: Add new IRQ flag IRQF_NO_SUSPEND
Bugzilla: 671147
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>

BZ671147

This is a partial backport of upstream commit
685fd0b4ea3f0f1d5385610b0d5b57775a8d5842, but the backport is even
more true to the summary then the original commit was, as we _only_
add thew new flag IRQF_NO_SUSPEND.

The purpose of the flag is to enable no-suspend behavior as the
name implies. Ideally we would redefine the IRQF_TIMER flag to
absorb this new flag, but then only test against the new flag when
checking if we should avoid suspending. That would allow us to stop
overloading the purpose of IRQF_TIMER. However, to preserve kABI we
should keep the old flags the same and just introduce the new one to
be used in the future.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>