perf: fix lock recursion
Patchwork-id: 32925
O-Subject: [RHEL6.1 PATCH 1/7] perf: fix lock recursion
Bugzilla: 593763
RH-Acked-by: Peter Zijlstra <pzijlstr@redhat.com>

In testing 'perf lock', I found that the system freezes up really fast and becomes
completely unresponsive. The issue is that when the locking tracepoints are
enabled we hit recursion in the perf layer, due to its use of locking. This
is solved by dropping additional traces that are issued by the perf layer.

The fix is based on the upstream use of:

extern int perf_swevent_get_recursion_context(void);
extern void perf_swevent_put_recursion_context(int rctx);

So before we call 'perf_tp_event()' to record a tracepoint event, we first call
'perf_swevent_get_recursion_context', and check if we are already in the
middle of recording a tracepoing. If we are not we call perf_tp_event(), and
signal that we have finished using perf_swevent_put_recursion_context().

Upstream, 'perf_swevent_put_recursion_context()' is included in the
'perf_tp_event()' function, so that callers do not need to perform
the put. However, this requires adding an additional argument to the
perf_tp_event() function. Since, perf_tp_event() is on the abi whitelist,
I've simply made the callers of perf_tp_event() explicitly call
perf_swevent_put_recursion_context.

Signed-off-by: Jason Baron <jbaron@redhat.com>