allow multiple versions of debuginfo packages to coexist
Patchwork-id: 32934
O-Subject: [RHEL6.1 PATCH v3] allow multiple versions of debuginfo packages to
	coexist
Bugzilla: 435941
RH-Acked-by: Don Zickus <dzickus@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=435941

Currently because of how the directories are created during build time, it's
not possible to install multiple debuginfo packages (source files would
conflict on /usr/src/kernels/2.6.32/). This patch fixes it.

This version reintroduces the %doc change: the path used with %doc macro
is where the file is located, not the installed location. It must be
changed or the file won't be found.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>