KVM: backport of SVM TSC init fixes
Patchwork-id: 32943
O-Subject: [RHEL6.1 01/11] KVM: backport of SVM TSC init fixes
Bugzilla: 651635
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Glauber Costa <glommer@redhat.com>

This is the first of a multi-part patch series which fixes numerous
problems with TSC and paravirt clock under KVM.  I start by addressing
a simple and clear bug, but the patches quickly become highly coupled,
so separating this into a larger sequence is not really possible.

Many had issues integrating backwards to RHEL because of a silly little
API change: the kvm_make_request() function was added instead of doing
bit manipulations on the request vector directly.  This required careful
hand adjustment.  Where noted, care should be taken with the review.

Unfortunately, this is quite tricky code, as the problem being solved -
creating a stable clock from unstable ones which don't stay in sync,
change frequency underneath you, hotplug in and out - and also need to
survive migration, suspend and host suspend - is rather difficult.

That said, the tricky bits have been upstream for quite a while, yes,
bugs were found, and I have integrated the needed bugfixes.  A couple
other fixes were necessary to get things to build on the RHEL6 branch,
which were very minor, like missing header files and a few #ifdefs.

I refrained from introducing the TSC trapping code or including the
KVM clock-synchronization changes, which are not yet upstream, but I
did include a very important improvement and bugfix which are not yet
included in the upstream code base.

The MD5 hashes are change numbers from Avi's kvm.git tree, not the
mainline kernel.

Up first,

BZ: 651635

58877679fd393d3ef71aa383031ac7817561463d
47008cd887c1836bcadda123ba73e1863de7a6c4

On reset, VMCB TSC should be set to zero.  Instead, code was setting
tsc_offset to zero, which passes through the underlying TSC.

The VMCB is reset whenever we receive a startup IPI, so Linux is setting
TSC back to zero happens very late in the boot process and destabilizing
the TSC.  Instead, just set TSC to zero once at VCPU creation time.

Signed-off-by: Zachary Amsden <zamsden@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>