Btrfs: fix split_leaf double split corner case
Patchwork-id: 32982
O-Subject: [RHEL6 PATCH 84/81] Btrfs: fix split_leaf double split corner case
Bugzilla: 663749
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

From: Chris Mason <chris.mason@oracle.com>

split_leaf was not properly balancing leaves when it was forced to
split a leaf twice.  This commit adds an extra push left and right
before forcing the double split in hopes of getting the slot where
we want to insert at either the start or end of the leaf.

If the extra pushes do work, then we are able to avoid splitting twice
and we keep the tree properly balanced.

Signed-off-by: Chris Mason <chris.mason@oracle.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>