export DSA_verify as a gpl symbol
Patchwork-id: 33006
O-Subject: [RHEL6 PATCH] crypto: export DSA_verify as a gpl symbol
Bugzilla: 673577
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

Bug #673577: FIPS certification requires exporting DSA_verify function
https://bugzilla.redhat.com/show_bug.cgi?id=673577

In order to pass FIPS CAVS testing and rely upon our kernel module signing
infrastructure, we need to be able to certify our DSA verification
implementation in the kernel. To do so, the DSA_verify function needs to be
exported, so that the test infrastructure bits can use it.

This is a RHEL-only patch, as upstream doesn't have our module signing
patches. It was added to RHEL5.4 as part of the fips effort (under
bugzilla #470111), and I swear I mentioned its exclusion when the modsign
patches for 6.0 were posted, but alas, this fell through the cracks. No
big deal, its a simple fix.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>