Fix NULL pointer dereference in cdrom driver
Patchwork-id: 33007
O-Subject: [RHEL6 PATCH] Fix NULL pointer dereference in cdrom driver
Bugzilla: 673567
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Bob Picco <bpicco@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=673567

This patch fixes a bug that Stratus stumbled across when we would test
thousands of failovers over a weekend.  Occasionally we would crash with
a NULL pointer dereference in cdrom_release().  This could happen if the
failover occurred at the exact moment the drive was being released.

I'm going to submit this upstream too.

 cdrom.c |    6 +++++-
 1 file changed, 5 insertions(+), 1 deletion(-)

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>