unmark gcm(aes) as fips_allowed
Patchwork-id: 33036
O-Subject: [RHEL6 PATCH] crypto: unmark gcm(aes) as fips_allowed
Bugzilla: 638133
RH-Acked-by: David S. Miller <davem@redhat.com>

Bugzilla #638133

This is a follow-up to a previously committed patch to fix the referenced
bugzilla. Its been determined that we don't have the capacity or
justification to certify gcm(aes) in fips mode for RHEL6, so it now needs
to be UN-marked as fips_allowed (along with reverting the fips_allowed
addition to ghash from the prior patch). And yes, this is a conscious
diversion from upstream, for business reasons.

Signed-off-by: Jarod Wilson <jarod@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>