mmapping a read only file on a gfs2 filesystem incorrectly acquires an exclusive glock
Patchwork-id: 33052
O-Subject: [RHEL 6.1] mmapping a read only file on a gfs2 filesystem
	incorrectly acquires an exclusive glock (bz #674286)
Bugzilla: 674286
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>
RH-Acked-by: Josef Bacik <josef@redhat.com>

This patch improves the test which is done before deciding to take a glock during
the mmap atime update code. The test now includes a check for the noatime mount
option. Other tests must be done under the glock to ensure that the information
in the inode upon which they are based is uptodate. This does not mean that we
need an exclusive lock though - a shared lock is perfectly ok as an exclusive lock
will be taken later if and when an atime update is actually required.

This affects BLAST users on GFS2. The patch is a direct copy of the upstream patch
which is currently in the -nmw GFS2 git tree (pending merge to Linus).

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>