kvm: keep track of which task is running a KVM vcpu
Patchwork-id: 33060
O-Subject: [RHEL 6.2 PATCH 10/12] kvm: keep track of which task is running a
	KVM vcpu
Bugzilla: 464276
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Avi Kivity <avi@redhat.com>

Fixes bug 464276

Upstream status: soon to go into Avi's tree

Keep track of which task is running a KVM vcpu.  This helps us
figure out later what task to wake up if we want to boost a
vcpu that got preempted.

Unfortunately there are no guarantees that the same task
always keeps the same vcpu, so we can only track the task
across a single "run" of the vcpu.

Signed-off-by: Rik van Riel <riel@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>