fix use-after-free in scsi_init_io()
Patchwork-id: 33077
O-Subject: [PATCH 1/1] scsi: fix use-after-free in scsi_init_io()
Bugzilla: 674064
RH-Acked-by: David Milburn <dmilburn@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for bz 674064.

I have not been able to replicate the problem (I checked
out the same box as in the bz), but from the oops it looks
like it is oopsing twords the end of scsi_init_io, right?
I could not get the exact same kernel or messing up gdb
because it was giving me bogus info for:
RIP: 0010:[<ffffffff8137c295>]  [<ffffffff8137c295>] scsi_init_io+0xc5/0x170
so I could not tell exactly where we oopsed.

Our kernel is missing the commit below, so I suspect this
is the problem we hit.

Either way though we need this patch since our kernel can
hit the problem, so I wanted to send it.

Patch has been tested while trying to recreate
the original problem.

>From the upstream commit:

commit 3a5c19c23db65a554f2e4f5df5f307c668277056
Author: James Bottomley <James.Bottomley@suse.de>
Date:   Mon Aug 16 10:06:26 2010 -0500

    [SCSI] fix use-after-free in scsi_init_io()

    we're using a pointer through a freed command to reset the request,
    which has shown up as an oops with slab poisoning:

    Reported-by: Tejun Heo <tj@kernel.org>
    Reported-by: Alexey Dobriyan <adobriyan@gmail.com>
    Signed-off-by: James Bottomley <James.Bottomley@suse.de>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>