bonding: interface doesn't issue IGMP report on slave interface during failover
Patchwork-id: 33176
O-Subject: [RHEL6.1 PATCH] bonding: interface doesn't issue IGMP report on
	slave interface during failover
Bugzilla: 640690
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>
RH-Acked-by: Amerigo Wang <amwang@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

RHBZ#: 640690
=======
https://bugzilla.redhat.com/show_bug.cgi?id=640690

Description:
=============
The first problem is that bonding uses the IGMP timers
following the specs, but when a failover happens, the backup
slave and the switch connected to this slave don't know about
the multicast traffic. This causes a traffic pause for up to
a minute. The patch fixes it by sending the membership immediately.

The second problem is that this customer is using VLAN on top
of bonding devices and when a link failure event happens,
no IGMP membership is sent. This happens because bonding
doesn't handle multicast groups on VLAN devices.

The third one is about losing the first packet, so it allows
sysadmins to adjust how many membership reports are sent when
a failover happens to make sure the switch is notified.

The forth one fixes a crash when trying to add or remove
a slave with the master interface down. It happens because
the workqueue is still uninitialized.

Upstream:
==========
commit ffa95ed50f9fb2d8faaa6bd73086a7056ea46a06
Author: Ben Hutchings <bhutchings@solarflare.com>
bonding: Change active slave quietly when bond is down

commit c2952c314b4fe61820ba8fd6c949eed636140d52
Author: Flavio Leitner <fleitner@redhat.com>
bonding: add retransmit membership reports tunable

commit e12b453904c54bbdc515778ff664d87a7f9473af
Author: Flavio Leitner <fleitner@redhat.com>
bonding: fix to rejoin multicast groups immediately

commit 5a37e8ca8536c47871d46c82211f399adf06fd44
Author: Flavio Leitner <fleitner@redhat.com>
bonding: rejoin multicast groups on VLANs

Test Build at:
===============
https://brewweb.devel.redhat.com/taskinfo?taskID=3095631

Testing Status:
===============
Unfortunately, the customer is able to validate the fix
just for RHEL-4 and RHEL-5 which are okay. However, they
have plans to start migrating to RHEL-6 any time soon,
so they have asked us to fix it in RHEL-6 as well.

I managed to reproduce the bug locally seeing the traffic
stopping after removing the cable from active slave with
-115.el6 kernel and managed to verify the fix using the
brew built kernel with the patch below as well.

Proposed Patch:
================
This patch is based on 2.6.32-114

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>