support for marking code as tech preview
Patchwork-id: 33177
O-Subject: [RHEL6 PATCH 2/2 V2] [misc] support for marking code as tech preview
Bugzilla: 645198
RH-Acked-by: Jon Masters <jcm@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=645198

RedHat kernel has support for lots of platforms and drivers.  Over time
we add features and drivers that aren't quite mature yet, but would like
to have customers test drive them publicly to help flush out the
remaining bugs.  We usually mark that code as 'tech preview' in some
documentation when we release the kernel in an update.

Over time it has become more difficult to remember what is marked as
'tech preview' and even more difficult to flag machines during
certification that accidentally use that code.

As a result our support organization has requested we add some code
that makes it easier to determine if a running kernel is using code
marked as 'tech preview'.

I implemented a simple hook into the kernel and module taint flag
mechanisms to accomplish this.  Therefore adding a simple line like

mark_tech_preview(msg, THIS_MODULE);

in something like the module_init of a function, will cause the
following to happen

- output in dmesg
- the taint bit set in /proc/sys/kernel/tainted
- the taint bit set in the module under /proc/modules

This allows automated tests scripts to easily detect if a 'tech preview'
piece of code has been executed.

However, note this function goes beyond just adding it into a module_init
function.  There are certain cases where the 'tech preview' code is
built-in.  In those cases, it would be expected to use appropriate code
to verify the 'tech preview' path and wrap the function call above with
something like:

if (special_path) // Or a PCI/CPU quirks
	mark_tech_preview("MySpecialDevice", NULL);

This can also be used for certain modules that are used across multiple
devices but we are only marking a certain device or class as 'tech preview'.
An example would be certain audio chips or networking chips.

Lastly, I included a documentation file that I hope could be updated when
people flag their code.  This would contain a current running list of code
that is marked as 'tech preview' for easy reference.

Feedback welcomed.

V2: fix compile error with undefined struct module in kernel.h

Signed-off-by: Don Zickus <dzickus@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>