kabi: Remove blk_queue_ordered to sync with official kABI
Patchwork-id: 33179
O-Subject: [RHEL6.1 PATCH 1/5] kabi: Remove blk_queue_ordered to sync with
	official kABI
Bugzilla:

The official kABI as released to customers does not contain the
symbol "blk_queue_ordered", which was removed prior to RHEL6.0 GA
in accordance with Bugzilla #636975. Therefore, first sync the
RHEL6 tree to be up to date with the official reference files.

Signed-off-by: Jon Masters <jcm@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>