kabi: Update Module.kabi reference files in sync with official kABI
Patchwork-id: 33182
O-Subject: [RHEL6.1 PATCH 2/5] kabi: Update Module.kabi reference files in sync
	with official kABI
Bugzilla:

The official kABI reference files (rhel6_kabi) and in-tree kernel
files need to be updated. This removes one problematic symbol that
was not present in the official kABI and adds checksums for two
others that were, and should be monitored by the kABI checker.

Signed-off-by: Jon Masters <jcm@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>