kabi: Add blk_queue_flush to the kABI
Patchwork-id: 33180
O-Subject: [RHEL6.1 PATCH 3/5] kabi: Add blk_queue_flush to the kABI
Bugzilla:

Prior to the release of RHEL6.0, we made some changes to kABI for
the benefit of future enablement of FUA patches in 6.1. As part of
that effort, Mike Snitzer suggested we add blk_queue_flush in 6.1
to replace functionality previously removed from kABI in 6.0.

Signed-off-by: Jon Masters <jcm@redhat.com>