xen: microcode: no backtrace on guest restore
Patchwork-id: 33190
O-Subject: [RHEL 6.1 PATCH] xen: microcode: no backtrace on guest restore
Bugzilla: 671161
RH-Acked-by: Anton Arapov <Anton@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

BZ671161

When save/restoring a xen guest the guest spews the following warning
to the console/dmesg on every restore if the microcode module was loaded

WARNING: at arch/x86/kernel/microcode_core.c:451 mc_sysdev_resume+0x67/0x70
<and a backtrace>

The code that produces this warning (below) has been around forever.
However, we haven't seen the problem before because the module was never
loaded. Recent changes to the microcode_ctl package have started loading
the module automatically on some platforms.

This patch from Ian Campbell is in route to upstream (currently in
the mm-tree). Here is some of what Ian says about it

<quote>
Since the real constraint here is that we are running on the CPU for
which we would like to load microcode (which in all practical
circumstances is CPU0) just check for that and return if we are resuming
a different CPU.

There is no danger of concurrent updates, even if we ignore the fact that
all but one CPUs are unplugged on native, because sysdev_resume() is single
threaded.
</quote>